@ fix: add try-catch to confirmDialog render + extract shared block-editor logic
HIGH #3 (admin-config.js):
- Wrap initial render() call in try-catch. If render throws (e.g.
  incompatible React version), cleanup() is called with false so the
  Promise resolves instead of hanging forever and leaking the DOM node.

HIGH #4 (PHP containers):
- Extract ~150 lines of duplicated block-editor-panel enqueue logic
  into HasBlockEditorPanel trait (resolveEditorContext,
  collectPanelSchemas, resolvePanelScriptAsset, panelAsset*, etc.)
- BlockEditorPanelContainer: 257 -> 62 lines
- MetaboxContainer: 399 -> 277 lines
- Net: -119 lines, no behavioral change
- Trait uses containerTypeForBlockPanel() abstract method to supply
  the containerType field (metabox vs block_editor_panel)
- MetaboxContainer keeps normalize_post_types() — used by classic
  metabox methods too

Build: pass  JS runtime: pass  PHPUnit: 98/98 pass  PHPStan: 0 errors
@
